+++ cpp06/ex00/Info.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Info.hpp                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jayi <jayi@student.42seoul.kr>             +#+  +:+       +#+        */
+/*   By: jayi <jayi@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/18 21:10:01 by jayi              #+#    #+#             */
-/*   Updated: 2022/04/18 22:52:49 by jayi             ###   ########.fr       */
+/*   Updated: 2022/04/22 14:14:38 by jayi             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,11 +21,11 @@ class Info {
     bool existsPoint;
     const std::string special;
     long long integer;
-    double decimal;
+    long double decimal;
 
     Info(void);
     Info(bool isMinus, bool existsPoint, const std::string special,
-         long long integer, double decimal);
+         long long integer, long double decimal);
     ~Info();
     Info &operator=(const Info &ref);
     Info(const Info &ref);
